Refactored entities into data and requests for request entities created for http requests and actual data entites
