+++ Reimbursement-Request-App/src/main/java/com/example/Reimbursement/Request/App/DAL/CategoryDAL/CategoryDALInterface.java
@@ -1,6 +1,6 @@
 package com.example.Reimbursement.Request.App.DAL.CategoryDAL;
 
-import com.example.Reimbursement.Request.App.Entities.Category;
+import com.example.Reimbursement.Request.App.Entities.Data.Category;
 
 import java.util.List;
