+++ Reimbursement-Request-App/src/main/java/com/example/Reimbursement/Request/App/DAL/RequestDAL/RequestDALInterface.java
@@ -1,6 +1,6 @@
 package com.example.Reimbursement.Request.App.DAL.RequestDAL;
 
-import com.example.Reimbursement.Request.App.Entities.Request;
+import com.example.Reimbursement.Request.App.Entities.Data.Request;
 
 import java.util.List;
