+++ Reimbursement-Request-App/src/main/java/com/example/Reimbursement/Request/App/DAL/SessionDAL/SessionDALImplementation.java
@@ -1,6 +1,6 @@
 package com.example.Reimbursement.Request.App.DAL.SessionDAL;
 
-import com.example.Reimbursement.Request.App.Entities.Session;
+import com.example.Reimbursement.Request.App.Entities.Data.Session;
 import com.example.Reimbursement.Request.App.Utilities.DatabaseConnection;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
