+++ Reimbursement-Request-App/src/main/java/com/example/Reimbursement/Request/App/DAL/SessionDAL/SessionDALInterface.java
@@ -1,6 +1,6 @@
 package com.example.Reimbursement.Request.App.DAL.SessionDAL;
 
-import com.example.Reimbursement.Request.App.Entities.Session;
+import com.example.Reimbursement.Request.App.Entities.Data.Session;
 
 public interface SessionDALInterface {
     Session addSession(Session session);
