+++ Reimbursement-Request-App/src/main/java/com/example/Reimbursement/Request/App/Entities/Requests/LoginRequest.java
@@ -1,4 +1,4 @@
-package com.example.Reimbursement.Request.App.Entities;
+package com.example.Reimbursement.Request.App.Entities.Requests;
 
 import lombok.Getter;
 import lombok.Setter;
