+++ Reimbursement-Request-App/src/test/java/com/example/Reimbursement/Request/App/DAL/CategoryDALTests.java
@@ -1,7 +1,7 @@
 package com.example.Reimbursement.Request.App.DAL;
 
 import com.example.Reimbursement.Request.App.DAL.CategoryDAL.CategoryDALImplementation;
-import com.example.Reimbursement.Request.App.Entities.Category;
+import com.example.Reimbursement.Request.App.Entities.Data.Category;
 import org.junit.Assert;
 import org.junit.Test;
