+++ Reimbursement-Request-App/src/test/java/com/example/Reimbursement/Request/App/DAL/SessionDALTests.java
@@ -1,7 +1,7 @@
 package com.example.Reimbursement.Request.App.DAL;
 
 import com.example.Reimbursement.Request.App.DAL.SessionDAL.SessionDALImplementation;
-import com.example.Reimbursement.Request.App.Entities.Session;
+import com.example.Reimbursement.Request.App.Entities.Data.Session;
 import org.junit.Assert;
 import org.junit.Test;
