+++ Reimbursement-Request-App/src/test/java/com/example/Reimbursement/Request/App/SAL/RequestSALTests.java
@@ -4,7 +4,7 @@ import com.example.Reimbursement.Request.App.DAL.CategoryDAL.CategoryDALImplemen
 import com.example.Reimbursement.Request.App.DAL.EmployeeDAL.EmployeeDALImplementation;
 import com.example.Reimbursement.Request.App.DAL.RequestDAL.RequestDALImplementation;
 import com.example.Reimbursement.Request.App.Entities.CustomExceptions.GeneralError;
-import com.example.Reimbursement.Request.App.Entities.Request;
+import com.example.Reimbursement.Request.App.Entities.Data.Request;
 import com.example.Reimbursement.Request.App.SAL.CategorySAL.CategorySALImplementation;
 import com.example.Reimbursement.Request.App.SAL.EmployeeSAL.EmployeeSALImplementation;
 import com.example.Reimbursement.Request.App.SAL.RequestSAL.RequestSALImplementation;
